% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDAT20.R
\name{BDAT20}
\alias{BDAT20}
\title{BDAT 2.0}
\usage{
BDAT20(speciesID, dbh, h, H1 = 1.3, D2 = 0, H2 = 0, hXwood, hInd,
  trunkInd, Az = 0, hTrunk, Zsh = 0, Zab = 0, sortInd = 0,
  FixLenDef = list(), NMaxFixLen, result = "Volume")
}
\arguments{
\item{speciesID}{... BDATBArtNr in BDAT}

\item{dbh}{D1 in BDAT}

\item{h}{H in BDAT}

\item{H1}{...}

\item{D2}{...}

\item{H2}{...}

\item{hXwood}{length of X-Wood; Hxh in BDAT}

\item{hInd}{height indicator; Hkz in BDAT}

\item{trunkInd}{trunk indicator; Skz in BDAT}

\item{Az}{What's the appropriate english name for Az?}

\item{hTrunk}{height of trunk; Hsh in BDAT}

\item{Zsh}{What's the appropriate english name for Zsh?}

\item{Zab}{english name for Zab?}

\item{sortInd}{sorting indicator; Sokz in BDAT}

\item{FixLenDef}{list with length 4; FixLngDef in BDAT}

\item{NMaxFixLen}{NMaxFixLng in BDAT}

\item{result}{maybe a character vector containing "Volume",
"DBH", "IFeh", "FixLength" or "NFixLength"; look at BDAT20 documentation}
}
\description{
Calculates volumes and sorts, useful for forest inventories.
}
