% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDATVolAB.R
\name{BDATVolAB}
\alias{BDATVolAB}
\title{BDAT Volume section A to B}
\usage{
BDATVolAB(speciesID, dbh, h, lh, uh, H1 = 1.3, D2 = 0, H2 = 0,
  SecLen = 0.1, IErr = 0, bark = T)
}
\arguments{
\item{speciesID}{here comes the description of speciesID...}

\item{dbh}{... and here the one of bhd}

\item{h}{...}

\item{lh}{...}

\item{uh}{...}

\item{H1}{maybe just link to BDAT20 for details of H1, D2, H2, IErr}

\item{D2}{...}

\item{H2}{...}

\item{IErr}{...}

\item{bark}{logical... is output attr(, "bark") really necessary???}

\item{SecLng}{section length, default is 0.1m...}
}
\value{
which values will be returned?
}
\description{
Calculates tree's volume beginning from section A to B.
Balk is included.
}
\details{
maybe some details? e.g. that function is vectorized and therefore one
can simply use a numeric vector as input (not only one number).
But using a numeric vector assumes all vectors need to be from  the same length.
}
\examples{
one example would be nice
}
