% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfNormal.R
\name{cdfNormal}
\alias{cdfNormal}
\title{test function for cdf}
\usage{
cdfNormal(q, mean = 0, sd = 1, type = c("Fortran", "R"))
}
\arguments{
\item{q}{vector of quantiles}

\item{mean}{vector of means, default is 0}

\item{sd}{vector of standard deviations, default is 1}

\item{type}{character vector: 'R' or 'Fortran'}
}
\description{
cdf implemented in R and fortran
}
\examples{
cdfNormal(1, type = "R")
cdfNormal(1, type = "Fortran")
}
