% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_height_of_d.R
\name{get_height_of_d}
\alias{get_height_of_d}
\title{Height of given diameter inside tree taper}
\usage{
get_height_of_d(sp, d, h, dx, H1 = 1.3, D2 = 0, H2 = 0, Hx = 0,
  IFeh = 0)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{dx}{diameter for which height is required}
}
\value{
height of diameter dx inside stem taper
}
\description{
this function calculates the height of a given diameter inside
a tree taper of given dimensions (dbh, h)
}
