% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vol.R
\name{get_vol}
\alias{get_vol}
\title{Volume of tree segment}
\usage{
get_vol(sp, d, h, lh, uh, H1 = 1.3, D2 = 0, H2 = 0, SekLng = 0.1,
  IFeh = 0, VolABmR = 0, VolABoR = 0, bark)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{lh}{lower height in tree of segment for which volume is required}

\item{uh}{upper height in tree of segment for which volume is required}

\item{bark}{boolean, if TRUE volume including bark is calculated}
}
\value{
volume with bark of given segment inside stem in cubic meter
}
\description{
this function calculates the volume of a tree segment given
 species, diameter, height (NOT YET: d03), beginning and end height
of the segment. One can choose if bark should be included or excluded via
the boolean parameter bark, which if set to TRUE includes bark.
}
