% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarkSize.R
\name{getBarkSize}
\alias{getBarkSize}
\title{double size of bark at height X}
\usage{
getBarkSize(speciesID, D1, H1 = 1.3, D2 = 0, H2 = 0, H, Hx)
}
\arguments{
\item{speciesID}{species code from BDAT}

\item{D1}{diameter at H1 (usually at breast height (dbh))}

\item{H1}{height 1}

\item{D2}{...}

\item{H2}{...}

\item{H}{total height}

\item{Hx}{height X}
}
\value{
double size of bark at height x
}
\description{
The bark's size at height X is doubled for a tree with dimensions <<dbh, H1, D2, H2, h>>.
}
\details{
...
}
