% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatHx.R
\name{getDatHX}
\alias{getDatHX}
\title{get diameter of given height inside tree taper w/ or w/o bark}
\usage{
getDatHX(sp, d, h, hx, H1 = 1.3, D2 = 0, H2 = 0, bark = TRUE)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) of tree}

\item{h}{height of tree}

\item{hx}{height for which diameter is required}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
taper form from the first NFI ('Bundeswaldinventur I') is used}

\item{H2}{according to D2, default is 0}

\item{bark}{boolean, if TRUE volume including bark is calculated (default)}
}
\value{
diameter at height \code{hx}
}
\description{
This function calculates the diameter of a given height inside
for a given tree of dimensions (dbh, h).
}
\examples{
getDatHx(sp=1, d=30, h=40, hx=20)
}
