% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDhVol.R
\name{getDhVol}
\alias{getDhVol}
\title{get coarse wood volume of tree}
\usage{
getDhVol(sp, d, h, H1 = 1.3, D2 = 0, H2 = 0, DHGrz = 7, HDHGrz = NULL,
  SekLng = 0.1, bark = TRUE)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
taper form from the first german NFI ('Bundeswaldinventur I') are used}

\item{H2}{measuring height of D2, default is 0}

\item{bark}{boolean, if TRUE volume including bark is calculated (default)}

\item{SegLen}{segment length for calculating the volume}
}
\value{
volume with or without bark of given segment AB in cubic meter
}
\description{
This function calculates the coarse wood volume of a tree given
species, diameter and height. One can choose if bark should be included or
excluded via the boolean parameter bark, which will include bark if it is set
to TRUE.
}
\details{
Strictly speaking \code{getDhVol} is a wrapper function. It loads the fortran
subroutine \code{BDATVOLDHMR} or \code{BDATVOLDHOR} (depending on whether or
not bark is set to TRUE or FALSE) within a call of \code{loadBDAT}.
}
\examples{
getDhVol(sp=1, d=30, h=40, bark = T)
}
