% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiameter.R
\name{getDiameter}
\alias{getDiameter}
\title{diameter at height X}
\usage{
getDiameter(speciesID, dbh, h, Hx, H1 = 1.3, D2 = 0, H2 = 0,
  bark = TRUE)
}
\arguments{
\item{speciesID}{... BDATBArtNr in BDAT}

\item{dbh}{... D1 in BDAT}

\item{h}{... H in BDAT}

\item{Hx}{...}

\item{H1}{...}

\item{D2}{...}

\item{H2}{...}
}
\value{
...
}
\description{
The diameter at height X is calculated for a tree with dimensions <<dbh, H1, D2, H2, h>>.
}
\details{
...
}
\examples{
...
}
