% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDoubleBark.R
\name{getDoubleBark}
\alias{getDoubleBark}
\title{get double bark thickness inside tree taper}
\usage{
getDoubleBark(sp, d, h, hx, H1 = 1.3, D2 = 0, H2 = 0)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) of tree}

\item{h}{height of tree}

\item{hx}{height for which diameter is required}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
taper form from the first NFI ('Bundeswaldinventur I') is used}

\item{H2}{according to D2, default is 0}
}
\value{
height of diameter \code{dx}
}
\description{
This function calculates bark's double thickness of a given
diameter inside for a given tree of dimensions (dbh, h).
}
\details{
Strictly speaking \code{getH} is a wrapper function. It loads the fortran
subroutine \code{BDATHXDX} within a call of \code{loadBDAT}.
}
\examples{
getDoubleBark(sp=1, d=30, h=40, hx=20)
}
