% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH.R
\name{getH}
\alias{getH}
\title{get height of given diameter inside tree taper w/ or w/o bark}
\usage{
getH(sp, d, h, dx, H1 = 1.3, D2 = 0, H2 = 0)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) of tree}

\item{h}{height of tree}

\item{dx}{diameter for which height is required}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
taper form from the first NFI ('Bundeswaldinventur I') is used}

\item{H2}{according to D2, default is 0}
}
\value{
height of diameter \code{dx}
}
\description{
This function calculates the height of a given diameter inside
for a given tree of dimensions (dbh, h).
}
\examples{
getH(sp=1, d=30, h=40, dx=20)
}
