% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeight.R
\name{getHeight}
\alias{getHeight}
\title{Height of given diameter inside tree taper}
\usage{
getHeight(speciesID, dbh, h, Dx, H1 = 1.3, D2 = 0, H2 = 0)
}
\arguments{
\item{speciesID}{species code from BDAT}

\item{dbh}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{Dx}{diameter for which height is required}

\item{H1}{...}

\item{D2}{...}

\item{H2}{...}
}
\value{
height of diameter dx inside stem taper
}
\description{
this function calculates the height of a given diameter inside
a tree taper of given dimensions (dbh, h)
}
