% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_height_of_d.R
\name{getHeightOfDiameter}
\alias{getHeightOfDiameter}
\title{Height of given diameter inside tree taper}
\usage{
getHeightOfDiameter(sp, d, h, dx, H1 = 1.3, D2 = 0, H2 = 0)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{dx}{diameter for which height is required}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
data from the first BWI (Bundeswaldinventur) are used}

\item{H2}{according to D2, default is 0}
}
\value{
height of diameter dx inside stem taper
}
\description{
This function calculates the height of a given diameter inside
a tree taper of given dimensions (dbh, h)
}
\examples{
getHeightOfDiameter(1, 30, 40, 20)
}
