% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSegVol.R
\name{getSegVol}
\alias{getSegVol}
\title{get volume of tree segment}
\usage{
getSegVol(sp, d, h, A, B, H1 = 1.3, D2 = 0, H2 = 0, SekLng = 0.1,
  VolABmR = 0, VolABoR = 0, bark = TRUE)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{A}{lower height A of segment AB for which volume is required}

\item{B}{upper height B of segment AB for which volume is required}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
taper form from the first german NFI ('Bundeswaldinventur I') are used}

\item{H2}{measuring height of D2, default is 0}

\item{bark}{boolean, if TRUE volume including bark is calculated (default)}

\item{SegLen}{segment length for calculating the volume}
}
\value{
volume with or without bark of given segment AB in cubic meter
}
\description{
This function calculates the volume of a tree segment given
 species, diameter, height, beginning and end height
of the segment. One can choose if bark should be included or excluded via
the boolean parameter bark, which will include bark if it is set to TRUE.
}
\examples{
getSegVol(sp=1, d=30, h=40, bark = T)
}
