% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_BDAT20.R
\name{getSort}
\alias{getSort}
\title{BDAT sort function}
\usage{
getSort(sp, d, h, hx, H1 = 1.3, D2 = 0, H2 = 0, Hkz = 0, Skz = 0,
  Az = 0, Hsh = 0.1, Zsh = 0, Zab = 7, Sokz = 1, FixLngDef = 0,
  NMaxFixLng = 0)
}
\arguments{
\item{sp}{...}

\item{d}{...}

\item{h}{...}
}
\description{
Vectorized BDAT sort function
}
