% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolume.R
\name{getVolume}
\alias{getVolume}
\title{Volume of tree segment}
\usage{
getVolume(sp, d, h, lh, uh, H1 = 1.3, D2 = 0, H2 = 0, SekLng = 0.1,
  VolABmR = 0, VolABoR = 0, bark = TRUE)
}
\arguments{
\item{sp}{species code from BDAT}

\item{d}{diameter in breast height (dbh) from tree}

\item{h}{height of tree}

\item{lh}{lower height in tree of segment for which volume is required}

\item{uh}{upper height in tree of segment for which volume is required}

\item{H1}{height where \code{d} was measured; default is 1.3[m] for dbh}

\item{D2}{diameter at second height \code{H2}; if \code{D2 = 0} (default)
data from the first BWI (Bundeswaldinventur) are used}

\item{H2}{according to D2, default is 0}

\item{bark}{boolean, if TRUE volume including bark is calculated (default)}

\item{SeqLen}{the tree is divided in \code{tree height / SeqLen}
compartiments so \code{SeqLen} influences the calculated precision}
}
\value{
volume with bark of given segment inside stem in cubic meter
}
\description{
This function calculates the volume of a tree segment given
 species, diameter, height (NOT YET: d03), beginning and end height
of the segment. One can choose if bark should be included or excluded via
the boolean parameter bark, which will include bark if it is set to TRUE.
}
\examples{
getVolume(1, 30, 40, 10, 20, bark = T)
}
