% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolumeShort.R
\name{getVolumenShort}
\alias{getVolumenShort}
\title{shortcut wrapper for \code{getVolume}}
\usage{
getVolumenShort(speciesID, dbh, h, lh, uh, H1 = 1.3, D2 = 0, H2 = 0,
  SecLen = 0.1, bark = T)
}
\arguments{
\item{speciesID}{...}

\item{dbh}{... and here the one of dbh}

\item{h}{...}

\item{lh}{...}

\item{uh}{...}

\item{H1}{maybe just link to BDAT20 for details of H1, D2, H2, IErr}

\item{D2}{...}

\item{H2}{...}

\item{bark}{logical... is output attr(, "bark") really necessary???}

\item{SecLng}{section length, default is 0.1m...}
}
\value{
which values will be returned?
}
\description{
Default values for H1, D2, H2
}
\details{
maybe some details? e.g. that function is vectorized and therefore one
can simply use a numeric vector as input (not only one number).
But using a numeric vector assumes all vectors need to be from  the same length.
}
\examples{
one example would be nice
}
