% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBDAT.R
\name{loadBDAT}
\alias{loadBDAT}
\title{load BDATPro}
\usage{
loadBDAT(fun, type = "expr")
}
\arguments{
\item{fun}{name of the function as character string(s) which shall be loaded
into global environment. See details.}

\item{type}{character vector; if type is \code{expr} (default), an expression
will be returned; if type is \code{global}, \code{fun} is loaded into global
environment}
}
\value{
Returns a fortran function and loads it into your global environment.
}
\description{
\code{loadBDAT} imports a dll file which contains a bunch of functions for
modelling tree taper given diameters and height of a tree.
}
\details{
Since this package contains the dll files compiled on Windows 32bit,
\code{loadBDAT()} access them directly. \cr
\code{fun} needs at least one of the following character strings to load a
fortran function of BDAT in your global environment: \cr
\itemize{
  \item 'BDATDORHX': diameter without bark in height x
  \item 'BDATDMRHX': diameter with bark in height x
  \item 'BDATHXDX': get height x for given diameter x
  \item 'BDATVOLDHOR': total volume with bark
  \item 'BDATVOLDHMR': total volume without bark
  \item 'BDATVOLABOR': volume without bark between height A und B
  \item 'BDATVOLABMR': volume with bark between height A und B
  \item 'BDATRINDE2HX': double bark thickness at height X
  \item 'V_BDAT20': vectorised sorting function inclusive fix length
}
Not yet implemented are the following BDAT-functions:
\itemize{
  \item 'BDATMwQ03BWI': Mittelwert, Streuung und Formquotienten-Verteilung
  \item 'BDATPctQ03BWI': Perzentilwert des Formquotienten
  \item 'FNBDATQ03VHDx': Volumenäquivalenter Formquotienten
  \item 'FNBDATEstQ032': Fortschreibung der Schaftform bei Folgeinventur
}
}
\examples{
loadBDAT(fun = 'BDATDORHX')
}
