% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBDAT.R
\name{loadBDAT}
\alias{loadBDAT}
\title{load BDATpro}
\usage{
loadBDAT(fun)
}
\arguments{
\item{fun}{name of the function as character string(s) which shall be loaded into global environment. see details}

\item{path}{path to BDATPRO_R.dll which need to be installed on your computer (use 32bit R version). see details}
}
\value{
Returns a fortran function translated in R and loads it into your global environment.
}
\description{
\code{loadBDAT} imports a dll file which contains a bunch of functions for
modelling different biomass parts.
}
\details{
\code{path} default is \code{"C:/Program Files (x86)/BDATPRO/R/BDATPRO_R.dll"} which is the link to installed BDAT on your FVA computer. \cr
\code{fun} needs at least one of the following character strings to load a fortran function of BDAT in your global environment: \cr
\itemize{
  \item 'BDATORHX': diameter without bark in height x
  \item 'BDATMRHX': diameter with bark in height x
  \item 'BDATVOLABOR': volume without bark between height A und B
  \item 'BDATVOLABMR': volume with bark between height A und B
  \item 'BDATRINDE2HX': double bark thickness at height X
  \item 'BDATHXDX': get height x for given diameter x
  \item 'V_BDAT20': sort function inclusive fix length
}
}
\examples{
loadBDAT(fun = 'BDATDORHX')
}
